+++ src/types/interfaces.ts
@@ -19,3 +19,13 @@ type Sourse = {
     id: string;
     name: string;
 };
+
+export interface SourseResponse {
+    id: string;
+    name: string;
+    description: string;
+    url: string;
+    category: string;
+    language: string;
+    country: string;
+}
